+++ CreationsForge.Assets/Files/AssetFileResolutionStatus.cs
@@ -1,0 +1,10 @@
+namespace CreationsForge.Assets.Files;
+
+public enum AssetFileResolutionStatus
+{
+    ResolvedLooseFile,
+    MissingAbsoluteFile,
+    MissingDataFolder,
+    MissingLooseFile,
+    ArchiveExtractionUnsupported
+}