+++ CreationsForge.Assets/Temp/IAssetTempFileSession.cs
@@ -1,0 +1,8 @@
+namespace CreationsForge.Assets.Temp;
+
+public interface IAssetTempFileSession : IDisposable
+{
+    string RootDirectory { get; }
+
+    string CreateExtractionDirectory(string scopeName);
+}