+++ CreationsForge.Core/Services/Interfaces/IAssetFileResolverService.cs
@@ -1,0 +1,9 @@
+using CreationsForge.Assets.Files;
+using CreationsForge.Core.DTOs.Assets;
+
+namespace CreationsForge.Core.Services.Interfaces;
+
+public interface IAssetFileResolverService
+{
+    AssetFileResolutionDTO ResolveAssetFile(AssetPreviewCandidateDTO candidate);
+}